Extracted winning check.
